update test refs to use public mock instruments
