+++ src/qcodes/tests/dataset/dond/test_do1d.py
@@ -12,14 +12,14 @@ from pytest import LogCaptureFixture
 from qcodes import config, validators
 from qcodes.dataset import do1d, new_experiment
 from qcodes.dataset.data_set import DataSet
-from qcodes.parameters import Parameter
-from qcodes.tests.dataset.conftest import ArrayshapedParam
-from qcodes.tests.instrument_mocks import (
+from qcodes.instrument_drivers.mock_instruments import (
     ArraySetPointParam,
     Multi2DSetPointParam,
     Multi2DSetPointParam2Sizes,
     MultiSetPointParam,
 )
+from qcodes.parameters import Parameter
+from qcodes.tests.dataset.conftest import ArrayshapedParam
 
 
 @pytest.mark.usefixtures("plot_close", "experiment")
