+++ src/qcodes/tests/parameter/test_val_mapping.py
@@ -3,8 +3,8 @@ from collections.abc import Generator
 import pytest
 
 import qcodes.validators as vals
+from qcodes.instrument_drivers.mock_instruments import DummyInstrument
 from qcodes.parameters import Parameter, create_on_off_val_mapping
-from qcodes.tests.instrument_mocks import DummyInstrument
 
 from .conftest import ParameterMemory
